Snapshot of kokkos-kernels.git from commit 7f8d1465a9f5158a42f94d8d43205eee225d82e3
From repository at git@github.com:kokkos/kokkos-kernels.git

At commit:
commit 7f8d1465a9f5158a42f94d8d43205eee225d82e3
Author: Nathan Ellingwood <ndellin@sandia.gov>
Date:   Wed Mar 26 12:25:13 2025 -0600

    Update master_history.txt

+++ packages/kokkos-kernels/sparse/unit_test/Test_Sparse_SortCrs.hpp
@@ -420,7 +420,7 @@ void testSortAndMerge(bool justGraph, int howExecSpecified, bool doStructInterfa
   EXPECT_EQ(goldEntries.size(), outEntries.extent(0));
   if (!justGraph) {
     EXPECT_EQ(goldValues.size(), outValues.extent(0));
-    EXPECT_EQ(goldValues.size(), output.nnz());
+    EXPECT_EQ(goldValues.size(), size_t(output.nnz()));
   }
   for (size_t i = 0; i < goldRowmap.size(); i++) EXPECT_EQ(goldRowmap[i], outRowmap(i));
   for (size_t i = 0; i < goldEntries.size(); i++) {
